refactor(sync_loop): Move `BlockValidator` and reduce scope
Do not expose `BlockValidator` outside of module `sync_loop`. Let the interface
to the sync loop, which is `SyncLoopHandle`, take a `Network` object instead.
The correct `BlockValidator` is then inferred from the passed `Network`.

This change reduces interface complexity.
